A new class is created for connection handlers
